SA - 3.1.0.52 - 4/15/2013 (stereo branch) 1) Implemented capability to move start/stop cursors in edit-segment and edit-boundaries mode using CTRL-ALT and SHIFT-ALT sequences with LEFT and RIGHT arrows.

+++ Src/SaExe/Process/Sa_p_spg.cpp
@@ -106,7 +106,7 @@ long CProcessSpectrogram::Exit(int nError)
 long CProcessSpectrogram::Process(void * pCaller, ISaDoc * pDoc, CSaView * pView, int nWidth, int /*nHeight*/, int nProgress, int nLevel)
 {
 
-    WORD wSmpSize = (WORD)(pDoc->GetSampleSize());
+    DWORD wSmpSize = (DWORD)(pDoc->GetSampleSize());
     // check canceled
     if (IsCanceled())
     {
